Rendering

+++ HighLo/src/Engine/Platform/Windows/WindowsFileSystem.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Can Karka and Albert Slepak. All rights reserved.
+// Copyright (c) 2021-2022 Can Karka and Albert Slepak. All rights reserved.
 
 #include "HighLoPch.h"
 #include "Engine/Core/FileSystem.h"
@@ -38,15 +38,117 @@ namespace highlo
             return ReadFileEx(file, buffer, (DWORD) size, &ol, FileIOCompleteRoutine);
         }
 
+        // @see: https://docs.microsoft.com/en-us/windows/win32/api/shellapi/nf-shellapi-shfileoperationw#return-value
+        static HLString TranslateWindowsShellOperationErrorCodeIntoHumanReadable(int32 errorCode)
+        {
+            switch (errorCode)
+            {
+                case 0x71:      return "The source and destination files are the same file.";
+                case 0x72:      return "Multiple file paths were specified in the source buffer, but only one destination file path.";
+                case 0x73:      return "Rename operation was specified but the destination path is a different directory. Use the move operation instead.";
+                case 0x74:      return "The source is a root directory, which cannot be moved or renamed.";
+                case 0x75:      return "The operation was canceled by the user, or silently canceled if the appropriate flags were supplied to SHFileOperation.";
+                case 0x76:      return "The destination is a subtree of the source.";
+                case 0x78:      return "Security settings denied access to the source.";
+                case 0x79:      return "The source or destination path exceeded or would exceed MAX_PATH.";
+                case 0x7A:      return "The operation involved multiple destination paths, which can fail in the case of a move operation.";
+                case 0x7C:      return "The path in the source or destination or both was invalid.";
+                case 0x7D:      return "The source and destination have the same parent folder.";
+                case 0x7E:      return "The destination path is an existing file.";
+                case 0x80:      return "The destination path is an existing folder.";
+                case 0x81:      return "The name of the file exceeds MAX_PATH.";
+                case 0x82:      return "The destination is a read-only CD-ROM, possibly unformatted.";
+                case 0x83:      return "The destination is a read-only DVD, possibly unformatted.";
+                case 0x84:      return "The destination is a writable CD-ROM, possibly unformatted.";
+                case 0x85:      return "The file involved in the operation is too large for the destination media or file system.";
+                case 0x86:      return "The source is a read-only CD-ROM, possibly unformatted.";
+                case 0x87:      return "The source is a read-only DVD, possibly unformatted.";
+                case 0x88:      return "The source is a writable CD-ROM, possibly unformatted.";
+                case 0xB7:      return "MAX_PATH was exceeded during the operation.";
+                case 0x402:     return "An unknown error occurred. This is typically due to an invalid path in the source or destination. This error does not occur on Windows Vista and later.";
+                case 0x10000:   return "An unspecified error occurred on the destination.";
+                case 0x10074:   return "Destination is a root directory and cannot be renamed.";
+            }
+
+            return "Unknown Error: " + HLString::ToString(errorCode);
+        }
+
+        // @see: https://docs.microsoft.com/de-de/windows/win32/debug/system-error-codes--0-499-
         static HLString TranslateErrorCodeIntoHumanReadable(DWORD errorCode)
         {
             switch (errorCode)
             {
+                case 0:
+                    return "success";
+
+                case 1:
+                    return "Wrong function";
+
+                case 2:
+                    return "File not found!";
+
+                case 3:
+                    return "The system could not find the file!";
+
+                case 4:
+                    return "The system could not open the file!";
+
+                case 5:
+                    return "Access denied!";
+
+                case 6:
+                    return "Invalid handle!";
+
+                case 7:
+                    return "NOT ENOUGH RAM!";
+
+                case 8:
+                    return "NOT ENOUGH RESOURCES";
+
+                case 9:
+                    return "INVALID ADDRESS";
+
+                case 10:
+                    return "Invalid format";
+
+                case 11:
+                    return "a program was loaded in a wrong format";
+
+                case 12:
+                    return "access key is invalid";
+
+                case 13:
+                    return "the data is invalid";
+
+                case 14:
+                    return "not enough space on the hard drive";
+
+                case 15:
+                    return "could not find the drive";
+
+                case 16:
+                    return "the directory could not been removed";
+
+                case 17:
+                    return "the file could not be copied onto another drive";
+
+                case 18:
+                    return "no files were found";
+
+                case 19:
+                    return "the medium is write secured";
+
+                case 20:
+                    return "the system could not find the device";
+
+                case 145:
+                    return "The directory is not empty!";
+
                 case 183:
-                    return "File could not be created - File already exists!";
+                    return "File already exists!";
             }
 
-            return "";
+            return "Unknown Error: " + HLString::ToString(errorCode);
         }
 
         static FileSystemPath GetFolderById(const KNOWNFOLDERID &id)
@@ -82,32 +184,31 @@ namespace highlo
         return (result != INVALID_FILE_ATTRIBUTES && (result & FILE_ATTRIBUTE_DIRECTORY));
     }
 
-    bool FileSystem::PathExists(const FileSystemPath &path)
-    {
-        HLString pathStr = path.String();
-        if (pathStr.IsEmpty())
-            return false;
-
-        DWORD result = GetFileAttributesW(pathStr.W_Str());
-        return (result != INVALID_FILE_ATTRIBUTES && (result & FILE_ATTRIBUTE_DIRECTORY));
-    }
-
     bool FileSystem::RemoveFile(const FileSystemPath &path)
     {
-        FileSystemWatcher::Get()->DisableWatchUntilNextAction();
+        return ::DeleteFileW(path.String().W_Str());
+    }
 
-        char *fp = (char*)*path.String();
-        fp[strlen(fp)] = '\0';
-        SHFILEOPSTRUCTA file_op;
-        file_op.hwnd = NULL;
-        file_op.wFunc = FO_DELETE;
-        file_op.pFrom = fp;
-        file_op.pTo = "";
-        file_op.fFlags = FOF_NOCONFIRMATION | FOF_NOERRORUI | FOF_SILENT;
-        file_op.fAnyOperationsAborted = false;
-        file_op.hNameMappings = 0;
-        file_op.lpszProgressTitle = "";
-        return SHFileOperationA(&file_op) == 0;
+    bool FileSystem::Copy(const FileSystemPath &dest, const FileSystemPath &src, bool overrideIfExist)
+    {
+        return ::CopyFileW(src.String().W_Str(), dest.String().W_Str(), !overrideIfExist);
+    }
+
+    bool FileSystem::CopyRecursive(const FileSystemPath &dest, const FileSystemPath &src)
+    {
+        HLString source = src.String();
+        if (!source.EndsWith("/"))
+        {
+            source = source + "\\*";
+        }
+
+        SHFILEOPSTRUCTW s = { 0 };
+        s.wFunc = FO_COPY;
+        s.pTo = dest.String().W_Str();
+        s.pFrom = source.W_Str();
+        s.fFlags = FOF_SILENT | FOF_NOCONFIRMMKDIR | FOF_NOCONFIRMATION | FOF_NOERRORUI | FOF_NO_UI;
+        int32 result = SHFileOperationW(&s);
+        return result == 0;
     }
 
     int64 FileSystem::GetFileSize(const FileSystemPath &path)
@@ -130,9 +231,37 @@ namespace highlo
         return success;
     }
 
-    bool FileSystem::RemoveFolder(const FileSystemPath &path)
+    bool FileSystem::RemoveFolder(const FileSystemPath &path, bool forceDelete)
     {
-        return RemoveDirectoryW(path.String().W_Str());
+        if (!forceDelete)
+        {
+            bool success = RemoveDirectoryW(path.String().W_Str());
+            if (!success)
+                HL_CORE_ERROR("WinAPI Error: {0}", *utils::TranslateErrorCodeIntoHumanReadable(GetLastError()));
+
+            return success;
+        }
+        else
+        {
+            // the user wants to force-delete the directory, without the limitation of only being able to remove an empty directory
+            // therefore we use the Windows Shell instead of the Win32 API
+
+            SHFILEOPSTRUCT fileOperation;
+            fileOperation.hwnd = NULL;
+            fileOperation.wFunc = FO_DELETE;
+            fileOperation.pFrom = path.String().W_Str();
+            fileOperation.pTo = L"";
+            fileOperation.fFlags = FOF_NOCONFIRMATION | FOF_NOERRORUI | FOF_SILENT;
+            fileOperation.fAnyOperationsAborted = false;
+            fileOperation.fFlags = 0;
+            fileOperation.hNameMappings = L"";
+
+            int32 status = SHFileOperationW(&fileOperation);
+            if (status != 0)
+                HL_CORE_ERROR("WinAPI Error: {0}", *utils::TranslateWindowsShellOperationErrorCodeIntoHumanReadable(status));
+
+            return status == 0;
+        }
     }
 
     HLString FileSystem::Rename(const FileSystemPath &path, const FileSystemPath &newName)
